codebase/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "valkyrie-training"
version = "0.1.0"
description = "High-performance training for Valkyrie (Longformer + S5) on TPU v4-32"
authors = [{name = "Valkyrie Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # JAX ecosystem
    "jax[tpu]>=0.4.20",
    "flax>=0.7.5",
    "optax>=0.1.7",
    "orbax-checkpoint>=0.4.0",
    
    # Data processing
    "datasets>=2.14.0",
    "transformers>=4.35.0",
    "tokenizers>=0.14.0",
    
    # Utilities
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    
    # TPU specific
    "libtpu-nightly",
    "jaxlib",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-xdist>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "S101", "PLR0913", "PLR0912", "PLR0915"]